Call peaks: SEACR

Call peaks from a bedGraph or bigWig file using SEACR (Sparse Enrichment Analysis for CUT&RUN).

call_peaks_seacr(
  bedgraph_path,
  control = 0.05,
  norm = TRUE,
  stringent = TRUE,
  outdir = tempdir(),
  outputfile = basename(tempfile(pattern = bedgraph_path, fileext = ".SEACR_peaks")),
  return_path = FALSE,
  remove_tmps = TRUE,
  nThread = 1,
  verbose = TRUE
)

Source

GitHub Issue: illegal byte sequence

Arguments

bedgraph_path

Field 1: Target data bedgraph file in UCSC bedgraph format that omits regions containing 0 signal.

control

Control (IgG) data bedgraph file to generate an empirical threshold for peak calling. Alternatively, a numeric threshold n between 0 and 1 returns the top n fraction of peaks based on total signal within peaks (default: 0.05).

norm

Field 3: "norm" (default: norm=TRUE) denotes normalization of control to target data, "non" (norm=FALSE) skips this behavior. "norm" is recommended unless experimental and control data are already rigorously normalized to each other (e.g. via spike-in).

stringent

Field 4:"relaxed" (stringent=FALSE) uses a total signal threshold between the knee and peak of the total signal curve, and corresponds to the “relaxed” mode described in the text, whereas “stringent” (default: stringent=TRUE) uses the peak of the curve, and corresponds to “stringent” mode.

outdir

Directory to store cutoff_analysis report and peak file in.

outputfile

Name of the peak output file (stored in BED format).

return_path

Whether to return the path to the saved peak file, or the peak data itself as a GRanges object.

remove_tmps

Remove temporary intermediate files generated by SEACR (i.e. )

nThread

When nThread>1, accelerates file importing and peak calling using multi-core parallelisation.

verbose

Print messages.

Value

GRanges or path to save peaks file.