Call consensus peaks: SEACR

Call consensus peaks by merging groups of BAM files with mergeBam and then calling peaks with call_peaks_seacr.

pooled_peaks_seacr(
  bam_files,
  outdir = tempdir(),
  g = NULL,
  overwrite = TRUE,
  nThread = 1,
  verbose = TRUE,
  ...
)

Arguments

bam_files

One or more BAM files.

outdir

Directory to save results to.

verbose

Print messages.

...

Arguments passed on to call_peaks_seacr

bedgraph_path

Field 1: Target data bedgraph file in UCSC bedgraph format that omits regions containing 0 signal.

control

Control (IgG) data bedgraph file to generate an empirical threshold for peak calling. Alternatively, a numeric threshold n between 0 and 1 returns the top n fraction of peaks based on total signal within peaks (default: 0.05).

norm

Field 3: "norm" (default: norm=TRUE) denotes normalization of control to target data, "non" (norm=FALSE) skips this behavior. "norm" is recommended unless experimental and control data are already rigorously normalized to each other (e.g. via spike-in).

stringent

Field 4:"relaxed" (stringent=FALSE) uses a total signal threshold between the knee and peak of the total signal curve, and corresponds to the “relaxed” mode described in the text, whereas “stringent” (default: stringent=TRUE) uses the peak of the curve, and corresponds to “stringent” mode.

remove_tmps

Remove temporary intermediate files generated by SEACR (i.e. )

nThread

When nThread>1, accelerates file importing and peak calling using multi-core parallelisation.

outputfile

Name of the peak output file (stored in BED format).

return_path

Whether to return the path to the saved peak file, or the peak data itself as a GRanges object.

Value

GRanges object.