Search metadata: AnnotationHub

Search metadata from AnnotationHub.

search_annotationhub(
  searches = construct_searches(),
  ah_id = NULL,
  output_category = NULL,
  output_type = NULL,
  file_format_type = NULL,
  biosample_id = NULL,
  biosample_type = NULL,
  biosample_name = NULL,
  dataset_biosample_summary = NULL,
  dataprovider = NULL,
  species = NULL,
  genome = NULL,
  description = NULL,
  rdataclass = NULL,
  sourcetype = NULL,
  partial_match = TRUE,
  peaks_only = FALSE,
  verbose = TRUE,
  ...
)

Source

https://biodatascience.github.io/compbio/bioc/anno.html Mike Love tutorial on accessing peak data from AnnotationHub.

Arguments

searches

Named list of regex queries.

ah_id

Search "ah_id" column.

output_category

Search "output_category" column.

output_type

Search "output_type" column.

file_format_type

Search "file_format_type" column.

biosample_id

Search "biosample_id" column.

biosample_type

Search "biosample_type" column.

biosample_name

Search "biosample_name" column.

dataset_biosample_summary

Search "dataset_biosample_summary" column.

dataprovider

Search "dataprovider" column.

species

Search "species" column.

genome

Search "genome" column.

description

Search "description" column.

rdataclass

Search "rdataclass" column.

sourcetype

Search "sourcetype" column.

partial_match

Return case-insensitive substring matches instead of exact matches only (default: TRUE).

peaks_only

Return pre-filtered metadata for peak data aligned to the GRCh38 human genome build.

verbose

Print messages.

...

Additional metadata columns to filter.

Value

data.table.

Examples

if (FALSE) { # \dontrun{
meta <- search_annotationhub(searches=construct_searches(keys="narrowpeak"),
                             biosample_type="brain",
                             peaks_only=TRUE)
} # }