Search metadata: ENCODE

Search metadata from the Encyclopedia of DNA Elements (ENCODE).

search_encode(
  assay = NULL,
  accession = NULL,
  file_accession = NULL,
  file_type = NULL,
  file_format = NULL,
  output_category = NULL,
  biosample_id = NULL,
  biosample_type = NULL,
  biosample_name = NULL,
  organism = NULL,
  dataset_type = NULL,
  assembly = NULL,
  genome_annotation = NULL,
  partial_match = TRUE,
  peaks_only = FALSE,
  verbose = TRUE,
  ...
)

Source

https://github.com/neurogenomics/EpiCompare/issues/43 Discussion on EpiCompare GitHub.

https://biodatascience.github.io/compbio/bioc/anno.html Mike Love tutorial on accessing peak data from AnnotationHub.

Arguments

assay

Search "assay" column.

accession

Search "accession" column.

file_accession

Search "file_accession" column.

file_type

Search "file_type" column.

file_format

Search "file_format" column.

output_category

Search "output_category" column.

biosample_id

Search "biosample_id" column.

biosample_type

Search "biosample_type" column.

biosample_name

Search "biosample_name" column.

organism

Search "organism" column.

dataset_type

Search "dataset_type" column.

assembly

Search "assembly" column.

genome_annotation

Search "genome_annotation" column.

partial_match

Return case-insensitive substring matches instead of exact matches only (default: TRUE).

peaks_only

Return pre-filtered metadata for peak data aligned to the GRCh38 human genome build.

verbose

Print messages.

...

Additional metadata columns to filter.

Value

data.table.

Examples

if (FALSE) { # \dontrun{
meta <- search_encode(assay = "chip-seq",
                      biosample_name = "K562",
                      file_type = "peak",
                      assembly = "GRCh38",
                      organism = "Homo sapiens",
                      partial_match = TRUE)
} # }