Search metadata: ENCODE

Search metadata from the Encyclopedia of DNA Elements (ENCODE).

search_roadmap(
  output_category = NULL,
  output_type = NULL,
  file_format_type = NULL,
  biosample_id = NULL,
  biosample_type = NULL,
  biosample_name = NULL,
  dataset_biosample_summary = NULL,
  dataprovider = NULL,
  species = NULL,
  genome = NULL,
  description = NULL,
  rdataclass = NULL,
  sourcetype = NULL,
  partial_match = TRUE,
  peaks_only = FALSE,
  verbose = TRUE,
  ...
)

Source

https://github.com/neurogenomics/EpiCompare/issues/43 Discussion on EpiCompare GitHub.

https://biodatascience.github.io/compbio/bioc/anno.html Mike Love tutorial on accessing peak data from AnnotationHub.

Arguments

output_category

Search "output_category" column.

output_type

Search "output_type" column.

file_format_type

Search "file_format_type" column.

biosample_id

Search "biosample_id" column.

biosample_type

Search "biosample_type" column.

biosample_name

Search "biosample_name" column.

dataset_biosample_summary

Search "dataset_biosample_summary" column.

dataprovider

Search "dataprovider" column.

species

Search "species" column.

genome

Search "genome" column.

description

Search "description" column.

rdataclass

Search "rdataclass" column.

sourcetype

Search "sourcetype" column.

partial_match

Return case-insensitive substring matches instead of exact matches only (default: TRUE).

peaks_only

Return pre-filtered metadata for peak data aligned to the GRCh38 human genome build.

verbose

Print messages.

...

Additional metdata columns to filter.

Value

data.table.

Examples

meta <- search_roadmap(biosample_type="brain",
                       file_format_type="narrowPeak",
                       genome="hg19",
                       peaks_only=TRUE)
#> Subsetting metadata to only peaks.
#> Filtering by: file_format_type
#> Filtering by: biosample_type
#> Filtering by: genome
#> Returning filtered metadata: 69 entries.